Move SAVED_STATUS and CANDIDATES_PER_PAGE to the scanning.lib and lower the number of candidate per page to speed up the scanning.
