+++ src/scanning/scanning.hooks.js
@@ -1,15 +1,16 @@
+import { useContext } from "react";
 import { useInfiniteQuery, useQuery } from "@tanstack/react-query";
 import { fetchAnnotationInfo, searchCandidates } from "./scanning.requests.js";
+import { CANDIDATES_PER_PAGE } from "./scanning.lib.js";
 import { fetchUserProfile } from "../onboarding/onboarding.lib.js";
-import { useContext } from "react";
 import { UserContext } from "../common/common.context.js";
-import { CANDIDATES_PER_PAGE, QUERY_KEYS } from "../common/common.lib.js";
+import { QUERY_KEYS } from "../common/common.lib.js";
 
 /**
  * @param {Object} props
  * @param {string} props.startDate - Only includes candidates that passed filters after this date.
  * @param {string} [props.endDate] - Only includes candidates that passed filters before this date.
- * @param {import("../common/common.lib.js").SavedStatus} props.savedStatus - The saved status of the candidates
+ * @param {import("./scanning.lib.js").SavedStatus} props.savedStatus - The saved status of the candidates
  * @param {number[]} props.groupIDs - The group IDs linked to the filter passed by the candidates.
  * @param {string} [props.queryID] - The query ID to filter candidates.
  * @returns {import("@tanstack/react-query").UseInfiniteQueryResult<import("@tanstack/react-query").InfiniteData<import("./scanning.requests.js").CandidateSearchResponse, unknown>, Error>}
